requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
Pillow>=10.0.0
python-magic>=0.4.27
firebase-admin>=6.2.0
boto3>=1.28.0
django-storages>=1.14.0
celery>=5.3.0
redis>=5.0.0
gunicorn>=21.0.0
